remove unsed photo carousel
