+++ src/components/Carousels/BaseCarousel/Cells/index.tsx
@@ -1,0 +1,2 @@
+export * from './PhotoCarouselCell';
+export { default as PhotoCarouselCell } from './PhotoCarouselCell';
