--- src/components/Carousels/PhotoCarousel/index.tsx
@@ -1,2 +1,0 @@
-export { default } from './PhotoCarousel';
-export * from './PhotoCarousel';
