+++ src/index.js
@@ -91,9 +91,6 @@ import { color, font, fontSize, mixins, spacing } from './styles';
 export { default as RelatedContentCard } from './components/Cards/RelatedContentCard';
 export * from './components/Cards/RelatedContentCard';
 
-export { default as PhotoCarousel } from './components/Carousels/PhotoCarousel';
-export * from './components/Carousels/PhotoCarousel';
-
 export { default as BylineList } from './components/BylineList';
 export * from './components/BylineList';
